refactor: theme-toggle
