+++ apps/status-page/src/components/status-page/floating-button.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ThemeToggle } from "@/components/theme-toggle";
+import { ThemeSelect } from "@/components/themes/theme-select";
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label";
 import {
@@ -310,7 +310,7 @@ export function FloatingButton({
               {IS_DEV ? (
                 <div className="space-y-2">
                   <Label htmlFor="theme">Theme</Label>
-                  <ThemeToggle id="theme" className="w-full" />
+                  <ThemeSelect id="theme" className="w-full" />
                 </div>
               ) : null}
               <div className="space-y-2">
